Move everything into test-harness
