+++ .gitignore
@@ -41,4 +41,4 @@ yarn-error.log*
 next-env.d.ts
 
 # test subset config
-packages/common/src/testUtil/testSubsetGrep.properties
+packages/test-harness/testSubsetGrep.properties
