+++ .vscode/tasks.json
@@ -50,7 +50,7 @@
       "label": "Ensure test subset file exists",
       "type": "npm",
       "script": "generate-test-subset-file-strict",
-      "path": "packages/cursorless-engine",
+      "path": "packages/test-harness",
       "presentation": {
         "reveal": "silent"
       },
@@ -60,7 +60,7 @@
       "label": "Generate test subset file",
       "type": "npm",
       "script": "generate-test-subset-file",
-      "path": "packages/cursorless-engine",
+      "path": "packages/test-harness",
       "presentation": {
         "reveal": "silent"
       },
