+++ packages/common/src/index.ts
@@ -1,7 +1,6 @@
 export * from "./cursorlessCommandIds";
 export * from "./testUtil/extractTargetedMarks";
 export { default as FakeIDE } from "./ide/fake/FakeIDE";
-export { runTestSubset, testSubsetGrepString } from "./testUtil/runTestSubset";
 export { Message } from "./ide/spy/SpyMessages";
 export { SpyIDERecordedValues } from "./ide/spy/SpyIDE";
 export { default as SpyIDE } from "./ide/spy/SpyIDE";
