+++ packages/test-harness/src/launchVscodeAndRunTests.ts
