+++ packages/test-harness/src/scripts/runUnitTestsOnly.ts
@@ -1,6 +1,6 @@
 /**
  * Runs all tests that don't have to be run within VSCode.
  */
-import { TestType, runAllTests } from "../util/runAllTests";
+import { TestType, runAllTests } from "../runAllTests";
 
 runAllTests(TestType.unit);
